create bq sanitizer
