+++ gcp_variant_transforms/libs/bigquery_row_generator.py
@@ -22,6 +22,7 @@ from typing import Any, Dict  # pylint: disable=unused-import
 
 from gcp_variant_transforms.beam_io import vcfio
 from gcp_variant_transforms.libs import bigquery_schema_descriptor  # pylint: disable=unused-import
+from gcp_variant_transforms.libs import bigquery_sanitizer
 from gcp_variant_transforms.libs import bigquery_util
 from gcp_variant_transforms.libs import processed_variant  # pylint: disable=unused-import
 from gcp_variant_transforms.libs import vcf_field_conflict_resolver  # pylint: disable=unused-import
@@ -35,6 +36,7 @@ _MAX_BIGQUERY_ROW_SIZE_BYTES = 10 * 1024 * 1024 - 10 * 1024
 # Number of bytes to add to the object size when concatenating calls (i.e.
 # to account for ", "). We use 5 bytes to be conservative.
 _JSON_CONCATENATION_OVERHEAD_BYTES = 5
+_BigQuerySchemaSanitizer = bigquery_sanitizer.SchemaSanitizer
 
 
 class BigQueryRowGenerator(object):
@@ -50,7 +52,7 @@ class BigQueryRowGenerator(object):
     # type: (...) -> None
     self._schema_descriptor = schema_descriptor
     self._conflict_resolver = conflict_resolver
-    self._bigquery_field_sanitizer = bigquery_util.BigQueryFieldSanitizer(
+    self._bigquery_field_sanitizer = bigquery_sanitizer.FieldSanitizer(
         null_numeric_value_replacement)
 
   def get_rows(self,
@@ -165,7 +167,7 @@ class BigQueryRowGenerator(object):
       alt_record = {bigquery_util.ColumnKeyConstants.ALTERNATE_BASES_ALT:
                     alt.alternate_bases}
       for key, data in alt.info.iteritems():
-        alt_record[bigquery_util.get_bigquery_sanitized_field_name(key)] = (
+        alt_record[_BigQuerySchemaSanitizer.get_sanitized_field_name(key)] = (
             data if key in alt.annotation_field_names else
             self._bigquery_field_sanitizer.get_sanitized_field(data))
       row[bigquery_util.ColumnKeyConstants.ALTERNATE_BASES].append(alt_record)
@@ -190,7 +192,7 @@ class BigQueryRowGenerator(object):
     # type: (...) -> (str, Any)
     if data is None:
       return None, None
-    field_name = bigquery_util.get_bigquery_sanitized_field_name(key)
+    field_name = _BigQuerySchemaSanitizer.get_sanitized_field_name(key)
     if not schema_descriptor.has_simple_field(field_name):
       raise ValueError('BigQuery schema has no such field: {}.\n'
                        'This can happen if the field is not defined in '
