+++ gcp_variant_transforms/vcf_to_bq.py
@@ -274,8 +274,8 @@ def run(argv=None):
              allow_incompatible_records=known_args.allow_incompatible_records,
              omit_empty_sample_calls=known_args.omit_empty_sample_calls,
              num_bigquery_write_shards=known_args.num_bigquery_write_shards,
-             null_numeric_value_replacement=
-             known_args.null_numeric_value_replacement))
+             null_numeric_value_replacement=(
+                 known_args.null_numeric_value_replacement)))
 
   result = pipeline.run()
   result.wait_until_finish()
